Add counter for WFP defense
This commit adds the counter that keeps track of the difference
between the outgoing and incoming packets in the client.
When the difference is positive, i.e., more outgoing traffic than
incoming traffic, loop cover packets should be sent in the real
queue. When the difference is negative, the client should send
loop cover packets.